fix(job): reorganize and fix margins
