+++ src/components/Work/ResumeJob/index.tsx
@@ -1,9 +1,16 @@
-import React, { FC, ReactNode, useState } from "react";
-import { Box, Button, Collapse, Flex, Heading } from "@chakra-ui/react";
+import React, { FC, ReactNode } from "react";
+import {
+  Box,
+  Button,
+  Collapse,
+  Flex,
+  Heading,
+  useDisclosure,
+} from "@chakra-ui/react";
 import { FaExpand } from "@react-icons/all-files/fa/FaExpand";
 import { FaWindowMinimize } from "@react-icons/all-files/fa/FaWindowMinimize";
-import { Job } from "./jobData";
-import JobContent from "./JobContent";
+import { Job } from "../jobData";
+import JobContent from "../JobContent";
 
 export interface ResumeJobProps {
   children?: ReactNode;
@@ -11,7 +18,7 @@ export interface ResumeJobProps {
 }
 
 const ResumeJob: FC<ResumeJobProps> = ({ children, job }) => {
-  const [isOpen, setIsOpen] = useState<boolean>(false);
+  const { isOpen, onToggle } = useDisclosure();
 
   return (
     <Box my={8}>
@@ -21,21 +28,25 @@ const ResumeJob: FC<ResumeJobProps> = ({ children, job }) => {
         px={6}
         py={4}
         justifyContent="space-between"
+        width={"64ch"}
       >
-        <Heading as="h3" fontSize="xl">
+        <Heading as="h3" fontSize="2xl">
           {job.name}
         </Heading>
 
-        <Button bg="transparent" onClick={() => setIsOpen(!isOpen)}>
+        <Button bg="transparent" onClick={onToggle}>
           {isOpen ? <FaWindowMinimize /> : <FaExpand />}
         </Button>
       </Flex>
 
-      <Collapse in={isOpen}>
+      <Collapse
+        // animateOpacity={false}
+        in={isOpen}
+        transition={{ enter: { height: { duration: 5 } } }}
+      >
         {children}
-        <Box bg="whiteAlpha.500" mx="auto" p={6} w="95%">
-          <JobContent job={job} />
-        </Box>
+
+        <JobContent job={job} />
       </Collapse>
     </Box>
   );
