+++ src/gatsby-browser.ts
@@ -1,0 +1,1 @@
+import "./styles.css";
