+++ src/styles.css
@@ -1,0 +1,7 @@
+html,
+body #___gatsby {
+  width: 100%;
+  height: 100%;
+  margin: 0;
+  padding: 0;
+}
